pyproject.toml
[project]
name = "pipecat-wavix"
version = "1.0.0"
description = "Wavix WebSocket serializer for Pipecat"
requires-python = ">=3.10"
readme = "README.md"
authors = [
    {name = "Wavix", email="support@wavix.com"}
]
dependencies = [
    "pipecat-ai[deepgram,runner,websocket]>=0.0.105",
    "loguru>=0.7.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "pyngrok>=7.5.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipecat_wavix"]
